Fix 62 moves
As of @a475ea5901, a basePower of `false` is a special value that
denotes that the move should always fail (unless modified). This
change broke 62 moves that currently use a basePower of `false` when
they appear to intend a basePower of 0. This commit changes the
basePower of these moves to 0.