Increase chance of /kill being logged
Despite the call to `rooms.lobby.logEntry`, use of /kill was not
actually being logged because the process terminated before the
log entry was written to the file. This commit introduces a 250
millisecond delay before killing the process in order to give the
log a chance to be written.

The reason we introduce a delay rather than synchronously wait for
the log to be written is to ensure that the server is actually killed
(e.g. in case of a problem writing the log).