v0.3.5 is the only version of node-sockjs known to work
After upgrading node-sockjs to v0.3.7 on the live server, there was an
issue where the `close` event was apparently not being fired on some
sockets, causing various problems.

The only version of ndoe-sockjs that we know for sure works is v0.3.5,
which is the version that had previously been in use on the main server
for a long time.