Add Stealth Rockless OU format
The OU council needs an OU format with no Stealth Rock to study the impact on the metagame this change would have, to be able to decide on a Stealth Rock ban in gen 6, which will shape the gen 6 OU metagame.
This metagame is not the same as a suspect test, so it cohabitates with regular OU and any ongoing suspect.
More info: http://www.smogon.com/forums/showpost.php?p=4725140&postcount=196