Fix Destiny Bond interaction with fwith future moves
Future moves were making both the user of them and
the Pokémon in the field to faint if the opponent used Destiny Bond
that turn.
Fix this by adding a property to future moves and making Destiny Bond
check it, since Destiny Bond does not work with future moves at all.