Grammatical mistake
The term "an" is only used when the word in-front is a 'consonant', any letter that is not a 'vowel' (a, e, i, o, u) whereas in this case, the term "an" is used/preferred over the term "a" although seeing examples above and below it is seemingly a mistake. This pull request is to fix that.