NEXT: Fix Echoed Voice/Bide | Add Phantom Force
Update Bide to use StallMove, due to it giving it's user Endure for the duration.
Update Echoed Voice, so that the first hit can be protected against, and the second cannot. Also make both hits considered Sound moves.
Update Phantom Force to be in line with other two turn charge moves.