Admins are able to override mutes and locks
Administrators shouldn't be able to be muted nor locked.
There is no instance in which an administrator should be legit punished.
In such a case, there's a worse problem than the administrator being able to be punished.

Additionally, if an administrator is muted or locked while on an alt, they can unmute themselves by using their admin account, despite being muted or locked.