Gen 1: Implement toxic counter properly
The toxic counter is a separate volatile.
It is set to 1 on burn, poison, and leech seed.
It starts on 1 on toxic, but gets increased by 1 every turn.
If the Pokémon gets both Toxic and leech seed, both get their damage increased.
If the Pokémon gets unstatused from Toxic but regains a damage dealing residual (brn, psn, leech seed, tox), the counter won't be reset until switch out.