Fix logic error in private room check for /invite
Fix possible crash by checking that roomauth exists in a secret room
with modjoin on before checking if the target user has permission to
join. This also fixes a really, really stupid logic error in the last
commit that makes it impossible for anyone that can't bypassall to get
invited anywhere.