Gen 1: Allow for minimum Attack Stat Exp and DVs
This change allows users to input less than perfect attack on their sets.
If they don't input any EVs or IVs, the game will max all of them.
Otherwise, it will check for the attack EVs and IVs set on the teambuilder to then calculate the proper stats, 30 atk being the minimum to max out the other stats.
A minimum of 1 DV is necessary to keep max HP.

This is just a temporary fix, once the gen 1 teambuilder is done we should just check for the proper limitations on acquiring Stat Exp.
Level of the Pokémon will be taken into account soon to calculate the min Atk Stat Exp as well.